+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 1.1.1 under development
 
-- Chg #43: Add `ext-zlib` to composer requirements (@DplusG)
-- Chg #40: Add support for `yiisoft/files` of version `^2.0` (@DplusG)
+- Bug #38: Drop `rotateByCopy` and make it the default (@DplusG)
+- Bug #43: Add `ext-zlib` to composer requirements (@DplusG)
+- Enh #40: Add support for `yiisoft/files` of version `^2.0` (@DplusG)
 
 ## 1.1.0 May 23, 2022